Import the existing code, more or less.
So far, I've mostly implemented a working proxy.  It seems to correctly pass back and forth requests and responses.  There are a couple of things not working; it doesn't maintain a session, which I suspect means I'm messing up the cookies, and I'm not sure it's correctly doing URL-rewriting on the way back.